+++ app/src/main/java/ca/pkay/rcloneexplorer/Services/DownloadService.java
@@ -5,9 +5,15 @@ import android.app.Notification;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
+import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
+import android.content.SharedPreferences;
+import android.net.wifi.WifiInfo;
+import android.net.wifi.WifiManager;
 import android.os.Build;
+import android.preference.PreferenceManager;
 import android.support.annotation.Nullable;
 import android.support.v4.app.NotificationCompat;
 import android.support.v4.app.NotificationManagerCompat;
@@ -21,6 +27,7 @@ import java.util.regex.Pattern;
 import ca.pkay.rcloneexplorer.BroadcastReceivers.DownloadCancelAction;
 import ca.pkay.rcloneexplorer.Items.FileItem;
 import ca.pkay.rcloneexplorer.Items.RemoteItem;
+import ca.pkay.rcloneexplorer.Log2File;
 import ca.pkay.rcloneexplorer.R;
 import ca.pkay.rcloneexplorer.Rclone;
 
@@ -37,7 +44,11 @@ public class DownloadService extends IntentService {
     private final int PERSISTENT_NOTIFICATION_ID = 167;
     private final int FAILED_DOWNLOAD_NOTIFICATION_ID = 138;
     private final int DOWNLOAD_FINISHED_NOTIFICATION_ID = 80;
+    private final int CONNECTIVITY_CHANGE_NOTIFICATION_ID = 235;
+    private boolean connectivityChanged;
+    private boolean transferOnWiFiOnly;
     private Rclone rclone;
+    private Log2File log2File;
     private Process currentProcess;
 
     /**
@@ -52,6 +63,14 @@ public class DownloadService extends IntentService {
         super.onCreate();
         setNotificationChannel();
         rclone = new Rclone(this);
+        log2File = new Log2File(this);
+
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        transferOnWiFiOnly = sharedPreferences.getBoolean(getString(R.string.pref_key_wifi_only_transfers), false);
+
+        if (transferOnWiFiOnly) {
+            registerBroadcastReceivers();
+        }
     }
 
     @Override
@@ -60,9 +79,18 @@ public class DownloadService extends IntentService {
             return;
         }
 
+        if (transferOnWiFiOnly && !checkWifiOnAndConnected()) {
+            showConnectivityChangedNotification();
+            stopSelf();
+            return;
+        }
+
         final FileItem downloadItem = intent.getParcelableExtra(DOWNLOAD_ITEM_ARG);
         final String downloadPath = intent.getStringExtra(DOWNLOAD_PATH_ARG);
         final RemoteItem remote = intent.getParcelableExtra(REMOTE_ARG);
+
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        Boolean isLoggingEnable = sharedPreferences.getBoolean(getString(R.string.pref_key_logs), false);
         
         Intent foregroundIntent = new Intent(this, DownloadService.class);
         PendingIntent pendingIntent = PendingIntent.getActivity(this, 0, foregroundIntent, 0);
@@ -91,22 +119,22 @@ public class DownloadService extends IntentService {
                     if (line.startsWith("Transferred:") && !line.matches("Transferred:\\s+\\d+$")) {
                         notificationBigText[0] = line;
                         notificationContent = line;
-                    }
-                    if (line.startsWith(" *")) {
+                    } else if (line.startsWith(" *")) {
                         String s = line.substring(2).trim();
                         notificationBigText[1] = s;
-                    }
-                    if (line.startsWith("Errors:")) {
+                    } else if (line.startsWith("Errors:")) {
                         notificationBigText[2] = line;
-                    }
-                    if (line.startsWith("Checks:")) {
+                    } else if (line.startsWith("Checks:")) {
                         notificationBigText[3] = line;
-                    }
-                    if (line.matches("Transferred:\\s+\\d+$")) {
+                    } else if (line.matches("Transferred:\\s+\\d+$")) {
                         notificationBigText[4] = line;
+                    } else if (isLoggingEnable && line.startsWith("ERROR :")){
+                        log2File.log(line);
                     }
 
-                    updateNotification(downloadItem, notificationContent, notificationBigText);
+                    if (transferOnWiFiOnly && !connectivityChanged) {
+                        updateNotification(downloadItem, notificationContent, notificationBigText);
+                    }
                 }
             } catch (IOException e) {
                 e.printStackTrace();
@@ -121,14 +149,47 @@ public class DownloadService extends IntentService {
 
         int notificationId = (int)System.currentTimeMillis();
 
-        if (currentProcess != null && currentProcess.exitValue() == 0) {
+        if (transferOnWiFiOnly && connectivityChanged) {
+            showConnectivityChangedNotification();
+        } else if (currentProcess != null && currentProcess.exitValue() == 0) {
             showDownloadFinishedNotification(notificationId, downloadItem.getName());
         } else {
-            rclone.logErrorOutput(currentProcess);
             showDownloadFailedNotification(notificationId, downloadItem.getName());
         }
 
         stopForeground(true);
+    }
+
+    private void registerBroadcastReceivers() {
+        IntentFilter intentFilter = new IntentFilter();
+        intentFilter.addAction(WifiManager.SUPPLICANT_CONNECTION_CHANGE_ACTION);
+        registerReceiver(connectivityChangeBroadcastReceiver, intentFilter);
+    }
+
+    private BroadcastReceiver connectivityChangeBroadcastReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            connectivityChanged = true;
+            stopSelf();
+        }
+    };
+
+    private boolean checkWifiOnAndConnected() {
+        WifiManager wifiMgr = (WifiManager) this.getSystemService(Context.WIFI_SERVICE);
+
+        if (wifiMgr == null) {
+            return false;
+        }
+
+        if (wifiMgr.isWifiEnabled()) { // Wi-Fi adapter is ON
+
+            WifiInfo wifiInfo = wifiMgr.getConnectionInfo();
+
+            return wifiInfo.getNetworkId() != -1;
+        }
+        else {
+            return false; // Wi-Fi adapter is OFF
+        }
     }
 
     private void updateNotification(FileItem downloadItem, String content, String[] bigTextArray) {
@@ -157,6 +218,17 @@ public class DownloadService extends IntentService {
 
         NotificationManagerCompat notificationManagerCompat = NotificationManagerCompat.from(this);
         notificationManagerCompat.notify(PERSISTENT_NOTIFICATION_ID, builder.build());
+    }
+
+    private void showConnectivityChangedNotification() {
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, CHANNEL_ID)
+                .setSmallIcon(android.R.drawable.stat_sys_warning)
+                .setContentTitle(getString(R.string.download_cancelled))
+                .setContentText(getString(R.string.wifi_connections_isnt_available))
+                .setPriority(NotificationCompat.PRIORITY_DEFAULT);
+
+        NotificationManagerCompat notificationManager = NotificationManagerCompat.from(this);
+        notificationManager.notify(CONNECTIVITY_CHANGE_NOTIFICATION_ID, builder.build());
     }
 
     private void showDownloadFinishedNotification(int notificationID, String contentText) {
@@ -225,6 +297,10 @@ public class DownloadService extends IntentService {
         super.onDestroy();
         if (currentProcess != null) {
             currentProcess.destroy();
+        }
+
+        if (transferOnWiFiOnly) {
+            unregisterReceiver(connectivityChangeBroadcastReceiver);
         }
     }