+++ app/src/main/java/ca/pkay/rcloneexplorer/Services/SyncService.java
@@ -5,13 +5,20 @@ import android.app.Notification;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
+import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
+import android.content.SharedPreferences;
+import android.net.wifi.WifiInfo;
+import android.net.wifi.WifiManager;
 import android.os.Build;
+import android.preference.PreferenceManager;
 import android.support.annotation.Nullable;
 import android.support.v4.app.NotificationCompat;
 import android.support.v4.app.NotificationManagerCompat;
 import android.support.v4.content.LocalBroadcastManager;
+import android.util.Log;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -19,6 +26,7 @@ import java.io.InputStreamReader;
 
 import ca.pkay.rcloneexplorer.BroadcastReceivers.SyncCancelAction;
 import ca.pkay.rcloneexplorer.Items.RemoteItem;
+import ca.pkay.rcloneexplorer.Log2File;
 import ca.pkay.rcloneexplorer.R;
 import ca.pkay.rcloneexplorer.Rclone;
 
@@ -33,7 +41,11 @@ public class SyncService extends IntentService {
     private final String CHANNEL_NAME = "Sync service";
     private final int PERSISTENT_NOTIFICATION_ID_FOR_SYNC = 162;
     private final int OPERATION_FAILED_NOTIFICATION_ID = 89;
+    private final int CONNECTIVITY_CHANGE_NOTIFICATION_ID = 462;
     private Rclone rclone;
+    private Log2File log2File;
+    private boolean connectivityChanged;
+    private boolean transferOnWiFiOnly;
     Process currentProcess;
 
     public SyncService() {
@@ -45,6 +57,14 @@ public class SyncService extends IntentService {
         super.onCreate();
         setNotificationChannel();
         rclone = new Rclone(this);
+        log2File = new Log2File(this);
+
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        transferOnWiFiOnly = sharedPreferences.getBoolean(getString(R.string.pref_key_wifi_only_transfers), false);
+
+        if (transferOnWiFiOnly) {
+            registerBroadcastReceivers();
+        }
     }
 
     @Override
@@ -53,10 +73,19 @@ public class SyncService extends IntentService {
             return;
         }
 
+        if (transferOnWiFiOnly && !checkWifiOnAndConnected()) {
+            showConnectivityChangedNotification();
+            stopSelf();
+            return;
+        }
+
         final RemoteItem remoteItem = intent.getParcelableExtra(REMOTE_ARG);
         final String remotePath = intent.getStringExtra(REMOTE_PATH_ARG);
         final String localPath = intent.getStringExtra(LOCAL_PATH_ARG);
         final int syncDirection = intent.getIntExtra(SYNC_DIRECTION_ARG, 1);
+
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        Boolean isLoggingEnable = sharedPreferences.getBoolean(getString(R.string.pref_key_logs), false);
 
         String title;
         int slashIndex = remotePath.lastIndexOf("/");
@@ -92,22 +121,22 @@ public class SyncService extends IntentService {
                     if (line.startsWith("Transferred:") && !line.matches("Transferred:\\s+\\d+$")) {
                         notificationBigText[0] = line;
                         notificationContent = line;
-                    }
-                    if (line.startsWith(" *")) {
+                    } else if (line.startsWith(" *")) {
                         String s = line.substring(2).trim();
                         notificationBigText[1] = s;
-                    }
-                    if (line.startsWith("Errors:")) {
+                    } else if (line.startsWith("Errors:")) {
                         notificationBigText[2] = line;
-                    }
-                    if (line.startsWith("Checks:")) {
+                    } else if (line.startsWith("Checks:")) {
                         notificationBigText[3] = line;
-                    }
-                    if (line.matches("Transferred:\\s+\\d+$")) {
+                    } else if (line.matches("Transferred:\\s+\\d+$")) {
                         notificationBigText[4] = line;
+                    } else if (isLoggingEnable && line.startsWith("ERROR :")){
+                        log2File.log(line);
                     }
 
-                    updateNotification(title, notificationContent, notificationBigText);
+                    if (transferOnWiFiOnly && !connectivityChanged) {
+                        updateNotification(title, notificationContent, notificationBigText);
+                    }
                 }
             } catch (IOException e) {
                 e.printStackTrace();
@@ -122,14 +151,47 @@ public class SyncService extends IntentService {
 
         sendUploadFinishedBroadcast(remoteItem.getName(), remotePath);
 
-        if (currentProcess == null || currentProcess.exitValue() != 0) {
-            rclone.logErrorOutput(currentProcess);
+        if (transferOnWiFiOnly && connectivityChanged) {
+            showConnectivityChangedNotification();
+        } else if (currentProcess == null || currentProcess.exitValue() != 0) {
             String errorTitle = "Sync operation failed";
             int notificationId = (int)System.currentTimeMillis();
             showFailedNotification(errorTitle, title, notificationId);
         }
 
         stopForeground(true);
+    }
+
+    private void registerBroadcastReceivers() {
+        IntentFilter intentFilter = new IntentFilter();
+        intentFilter.addAction(WifiManager.SUPPLICANT_CONNECTION_CHANGE_ACTION);
+        registerReceiver(connectivityChangeBroadcastReceiver, intentFilter);
+    }
+
+    private BroadcastReceiver connectivityChangeBroadcastReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            connectivityChanged = true;
+            stopSelf();
+        }
+    };
+
+    private boolean checkWifiOnAndConnected() {
+        WifiManager wifiMgr = (WifiManager) this.getSystemService(Context.WIFI_SERVICE);
+
+        if (wifiMgr == null) {
+            return false;
+        }
+
+        if (wifiMgr.isWifiEnabled()) { // Wi-Fi adapter is ON
+
+            WifiInfo wifiInfo = wifiMgr.getConnectionInfo();
+
+            return wifiInfo.getNetworkId() != -1;
+        }
+        else {
+            return false; // Wi-Fi adapter is OFF
+        }
     }
 
     private void updateNotification(String title, String content, String[] bigTextArray) {
@@ -166,6 +228,10 @@ public class SyncService extends IntentService {
         if (currentProcess != null) {
             currentProcess.destroy();
         }
+
+        if (transferOnWiFiOnly) {
+            unregisterReceiver(connectivityChangeBroadcastReceiver);
+        }
     }
 
     private void sendUploadFinishedBroadcast(String remote, String path) {
@@ -189,6 +255,17 @@ public class SyncService extends IntentService {
         NotificationManagerCompat notificationManager = NotificationManagerCompat.from(this);
         notificationManager.notify(notificationId, builder.build());
 
+    }
+
+    private void showConnectivityChangedNotification() {
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, CHANNEL_ID)
+                .setSmallIcon(android.R.drawable.stat_sys_warning)
+                .setContentTitle(getString(R.string.sync_cancelled))
+                .setContentText(getString(R.string.wifi_connections_isnt_available))
+                .setPriority(NotificationCompat.PRIORITY_DEFAULT);
+
+        NotificationManagerCompat notificationManager = NotificationManagerCompat.from(this);
+        notificationManager.notify(CONNECTIVITY_CHANGE_NOTIFICATION_ID, builder.build());
     }
 
     private void createSummaryNotificationForFailed() {